requirements.txt
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cryptography==42.0.8
fiona==1.9.6
Flask==3.0.3
Flask-Cors==4.0.1
geopandas==0.14.4
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
networkx==3.3
numpy==1.26.4
osmnx==1.9.3
packaging==24.0
pandas==2.2.2
pycparser==2.22
PyMySQL==1.1.1
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
scikit-learn==1.5.0
scipy==1.13.1
shapely==2.0.4
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
